requirements.txt
discord
python-decouple
chatterbot

google-api-python-client
oauth2client
pandas
CronTab